+++ config.go
@@ -1,6 +1,11 @@
 package logger
 
-import "go.uber.org/zap"
+import (
+	"time"
+
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+)
 
 type FieldPair []string
 
@@ -42,23 +47,27 @@ type Config struct {
 
 func NewProductionConfig(fields ...FieldPair) *Config {
 	return &Config{
-		Level:         InfoLevel,
-		Development:   false,
-		Encoding:      "json",
-		OutputPaths:   []string{"stdout"},
-		CallerSkip:    2,
-		InitialFields: genInitialFields(fields),
+		Level:             InfoLevel,
+		Development:       false,
+		Encoding:          "json",
+		OutputPaths:       []string{"stdout"},
+		CallerSkip:        2,
+		DisableStacktrace: false,
+		InitialFields:     genInitialFields(fields),
 	}
 }
 
 func NewDevelopmentConfig(fields ...FieldPair) *Config {
 	return &Config{
-		Level:         DebugLevel,
-		Development:   true,
-		Encoding:      "console",
-		OutputPaths:   []string{"stdout"},
-		CallerSkip:    2,
-		InitialFields: genInitialFields(fields),
+		Level:             DebugLevel,
+		Development:       true,
+		ShortTime:         true,
+		EnableColor:       true,
+		Encoding:          "console",
+		OutputPaths:       []string{"stdout"},
+		CallerSkip:        2,
+		DisableStacktrace: true,
+		InitialFields:     genInitialFields(fields),
 	}
 }
 
@@ -72,4 +81,69 @@ func genInitialFields(args []FieldPair) map[string]interface{} {
 		fields[f[0]] = f[1]
 	}
 	return fields
+}
+
+func (c *Config) buildZapConfig() {
+	encoderConfig := c.newCustomEncoderConfig()
+
+	zapConfig := &zap.Config{
+		Level:             zap.NewAtomicLevelAt(zapcore.Level(c.Level)),
+		Development:       c.Development,
+		DisableCaller:     c.DisableCaller,
+		DisableStacktrace: c.DisableStacktrace,
+		Sampling:          &zap.SamplingConfig{Initial: 100, Thereafter: 100},
+		Encoding:          c.Encoding,
+		EncoderConfig:     encoderConfig,
+		OutputPaths:       c.OutputPaths,
+		InitialFields:     c.InitialFields,
+	}
+	c.zapConfig = zapConfig
+}
+
+func (c *Config) newCustomEncoderConfig() zapcore.EncoderConfig {
+	encodeLevel := zapcore.LowercaseLevelEncoder
+	if c.EnableColor {
+		encodeLevel = zapcore.LowercaseColorLevelEncoder
+	}
+	encodeTime := zapcore.ISO8601TimeEncoder
+	if c.ShortTime {
+		encodeTime = func(t time.Time, enc zapcore.PrimitiveArrayEncoder) {
+			type appendTimeEncoder interface {
+				AppendTimeLayout(time.Time, string)
+			}
+			layout := "2006-01-02 15:04:05"
+			if enc, ok := enc.(appendTimeEncoder); ok {
+				enc.AppendTimeLayout(t, layout)
+				return
+			}
+			enc.AppendString(t.Format(layout))
+		}
+	}
+	return zapcore.EncoderConfig{
+		TimeKey:        "ts",
+		LevelKey:       "level",
+		NameKey:        "logger",
+		CallerKey:      "caller",
+		MessageKey:     "msg",
+		StacktraceKey:  "stacktrace",
+		LineEnding:     zapcore.DefaultLineEnding,
+		EncodeLevel:    encodeLevel,
+		EncodeTime:     encodeTime,
+		EncodeDuration: zapcore.SecondsDurationEncoder,
+		EncodeCaller:   zapcore.ShortCallerEncoder,
+	}
+}
+
+func (c *Config) clone() *Config {
+	cloned := *c
+	cloned.OutputPaths = make([]string, len(c.OutputPaths))
+	copy(cloned.OutputPaths, c.OutputPaths)
+	cloned.InitialFields = make(map[string]interface{})
+	for k, v := range c.InitialFields {
+		cloned.InitialFields[k] = v
+	}
+	if cloned.zapConfig != nil {
+		cloned.buildZapConfig()
+	}
+	return &cloned
 }